added - perimeter verify - area verify refactored some steps classes

+++ src/main/java/models/DigitResult.java
@@ -8,6 +8,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class DigitResult {
 
     @JsonProperty("result")
-    public int value;
+    public double value;
 
 }
